Create bottom navbar with passage navigation
- Create new BottomNav component (Footer)
- Move previous button, passage title, and next button to bottom
- Remove navigation controls from MyHeader
- MyHeader now only contains: 🔍 ➕ 🔊 (centered) | ☰ (right)
- BottomNav contains: ← 'Joh 1' → (centered)
- Passage title in bottom nav opens BibleSelector
- Clean separation: top for actions, bottom for navigation
- Remove unused imports from MyHeader
- Add footer to AppShell in App.tsx
- Build passes successfully
